requirements.txt
apache-airflow[postgres]==1.10.9
boto3